Add files via upload
Our team structure is a little abnormal, I, Mark Sowders, am the only person with a lot of Unity experience, so I am the only one with the skills to understand Unity well enough to make the UML and SD correctly(I don't mean right syntax, I mean right structure). So I build the UML and SD's and then we all had a team meeting in which I explained the structure. I know we are all supposed to work together, and Mark building everything might not look like that is what is going on, but trust me this is the best way to manage this project, and again we did all come together and I taught them what the heck is going on. This was not actually the initial plan, to demonstrate why the "Mark do it then teach it" plan was put in place, look at the extra SD submitted by Julian, you'll quickly see the difference.

Full transparency, after the meeting I sent the SD to each person to commit.